refactor: 디자인 패턴 수정
atomic 디자인 패턴에 맞게 코드 수정
기존 페이지에 컴포넌트를 맞게 작성
