+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import '@/styles/globals.css';
 import type { Metadata } from 'next';
 import Link from 'next/link';
-import Modal from '@/components/molecules/Modal';
 import { Inter } from '@next/font/google';
 
 const inter = Inter({
